Combine all timeseries tests in core/test_timeseries.py
